Implements increaseBudget method

+++ src/main/java/pl/com/dariusz/giza/financeTracker/repositories/BudgetsRepository.java
@@ -3,5 +3,5 @@ package pl.com.dariusz.giza.financeTracker.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.com.dariusz.giza.financeTracker.domain.budgets.Budget;
 
-public interface BudgetsRepository extends JpaRepository<Budget, Integer> {
+public interface BudgetsRepository extends JpaRepository<Budget, Long> {
 }
